add generic trait for FindById that throws exception if none found
